+++ src/battleorder_calc.c
@@ -7,7 +7,7 @@ u8 hp_condition(u8 bank, u8 percent);
 u8 get_item_effect(u8 bank, u8 check_negating_effects);
 u8 has_ability_effect(u8 bank, u8 mold_breaker);
 
-bool check_ability(u8 bank, u8 ability)
+bool check_ability(u8 bank, u16 ability)
 {
     if (has_ability_effect(bank, 0) && gBankAbilities[bank] == ability)
         return 1;
