+++ src/item_effects.c
@@ -6,7 +6,7 @@ u8 cant_become_burned(u8 bank, u8 self_inflicted);
 u8 cant_become_freezed(u8 bank, u8 self_inflicted);
 u8 cant_become_paralyzed(u8 bank, u8 self_inflicted);
 u8 cant_poison(u8 atk_bank, u8 def_bank, u8 self_inflicted);
-bool check_ability(u8 bank, u8 ability);
+bool check_ability(u8 bank, u16 ability);
 u8 get_item_effect(u8 bank, bool check_negating_effects);
 bool has_ability_effect(u8 bank, u8 mold_breaker);
 bool is_bank_present(u32 bank);
